Initial commit.
This project is largely an experiment to try out an organizational
approach to my code. The approach is to give most objects a "manager",
which mediates access and bundles convenience methods related to that
object.

I'd hit a wall with a previous project, probably due to some small
misunderstanding, where my attempts to encapsulate symbols into
different contexts resulted in "splits" of context: the control object
that event listeners were modifying was not the control object that
entities were querying. The map object that entities were updating was
not the map object that the renderer was using the draw. It was fixable,
but only with a lot of rework, and the approach used at the time was
likely to cause more splits in the future.

I believe that this approach is "too complex", and wouldn't necessarily
want to work on a real project that uses it, but it should provide an
interesting study piece for me to find a simpler approach that can
accomplish similar goals.

At the moment, there's a problem here requiring a somewhat significant
refactor: the map, entity, and player managers have very muddy scopes
of concern. The Map Manager is handling drawing and loading "the maps",
which are essentially a spec for a whole game, and the Player and
Entity managers are basically responsible for building representations
based on what the Map manager will tell them, then updating it when
they have updates.

Instead, I think it makes sense to have the Entity manager be responsible
for constructing the entity list out of the spec, and then registering
entities with a Renderer, and sending updates to that list. The Renderer
would then handle drawing its list of registered entities. A MapManager
would still be present, but just for querying the world's maps and their
limits, offering clamping to them, etc. The Player Manager will also still
exist, but pretty much entirely just operate on a special entity owned
by the Entity Manager.

Making the change from the former setup to the latter certainly justifies
a savepoint, so here we are.