Create RenderManager and use it.
This actually also necessitated a change in the data-model in a
great way -- rather than store the x_position in the entity records
and the source_x, source_width, dest_x, and dest_width in the map's
records, and performing a translation between the two when doing each
update, I should just have the record-sets be shared and use x_position
and x_scale for tile/entity reckoning. The resource can keep track of
its own source_x and source_width, as that relates solely to the image
it is being drawn from, and the tile/entity can keep track of where to
draw that rect to and how large to draw it.

This is only half of the battle -- the entity manager should be doing
the updates/dispatches and owning the models, rather than the render
manager. The entity manager shouldn't have to know about the render
manager at all; the frame manager (which may be combined with the
render manager very soon) should ask the entity manager for the entity
list and then act upon it. The player manager likewise shouldn't be
speaking directly to the render manager.