Combine FrameManager and RenderManager.
The responsibilities here are so similar that they should be taken
care of by one thing, so now they are. The new RenderManager takes
care of both the list of objects to draw and the frame logic.

There's a second ordering conflict in the managers now, where the
PlayerManager now also needs access to the RenderManager before it
has been instantiated. For the time being I've written a convenience
method to set the RenderManager within the PlayerManager until the
conflict is resolved -- resolving both should be as easy as making the
PlayerManager not know about the EntityManager or the RenderManager.

There's still some uncomfortable overlap between EntityManager and
PlayerManager that I'll have to resolve in the future!