Justify EntityManager's existence.
This commit makes the entity manager the provider of the draw-list,
which it forms pretty much by doing the work the render manager used
to do and just passing that list off to the render manager.

Not only that, this commit fixes up/calls the "update" method from
individual entities, so that they can mutate their state each frame,
and adds one test entity (the moving coin) to ensure that code is
working.

There's also some now-dead-code removed, and the comment that outlined
this plan is dropped, along with some other small cleanup changes.