Insert player layer ephemerally and cleanups.
The Player is no longer specified inside of a Map, but the map does
specify a player_layer for the player to be drawn at -- they will
be inserted at that layer index and the layer previously residing at
that index will be drawn above the player, as will all higher-indexed
layers above that.

This commit also (very sloppily) has a *ton* of cruft-cleaning: there
is no longer any passing of the raw config into the managers, only the
config manager. It's also no longer kept around by any manager; they use
the config and then forget about it. There are fewer console.logs, fewer
comments, no debug statements, and some of the common bits of code have
been made to look the same.