Implement basic collision detection.
Calculates the midpoint of tiles as rectangles, then says collision
is happening if the distance between two rect midpoints is less than
or equal to half of the largest any of the rects' axes.

Other subordinate or tangential changes in this commit:

- add physics manager for collision detection
- re-add "x_size" and "y_size" to tiles for collision
- re-id the tiles in the sample spec to be more semantically meaningful
- add entity manager code to calculate collision for one entity against
  its entity list
- add some convenience calls to entity and player managers for object access and modification
- adopt paradigm where entity manager passes itself to each entity during
  that entity's update() call, to give the entity access to other entities
  (including the player), and the physics engine.
- add test code in one coin entity to collide with the player, increase the player's score, and deactivate the entity
- modify render manager to respect an object being deactivated
- a few ' -> " and indentation fixes.

A little sloppy to do so much in a commit, but I'd rather just move
forward than sit around splitting it into a bunch of tiny discrete chunks.
My apologies to myself in the future when I try to figure out what the
heck I was thinking and wish I'd made the commits logical units of work.