Fix stale-player bug.
This was interesting. The player inserted into the entity-list was
actually just a small dictionary representing the player at a snap-
shot in time, which I hadn't noticed because the entity-list was
secretly being rebuilt *every single frame*.

I got most of the way into writing the multiple-maps code and fixed
some consistency issues* with how map-ids are handled, and this unknown
bug where the stale_entities() check had been *continuously failing*
got fixed! Suddenly the map was only being rebuilt when it was
actually needed, rather than every single frame. But I didn't even
know it was being built every single frame. The only impact was that
the player seemed to suddenly only be drawn when the map changed.

So, two fixes:
- put a real reference to the player into the entity list so it is
  updating live
- put a console.log into the stale entity code so that, if it starts
  failing 100% of the time again, it'll be pretty visible.

* that consistency issue was fixed by making the ids consistent, but
they are ugly array indices for now and I have yet to make them real
ids.