Implement multiple-map support.
A lot of the legwork was done on this previously -- the MapManager
is already dealing with an array rather than a single map, and both
it and the EntityManager both work with/save a current_map_id and work
with ids.

There were some consistency issues and bugs around the way the ids
were being used which have been fixed in this code, but not into the
best possible form -- they are array indices now rather than proper
unique ids stored in each map object. That fix will come later.

This change also adds a small limiter on how quickly maps can be changed
for the sake of my own testing, a control for cycling maps (the M key
at present), and a second map so I could test the cycling.