Fix quadtree entity movement.
Entities weren't actually updating their position within the quadtree,
and they only get rendered when they're stored in the region of the
quadtree which the camera is querying -- they all got left at their
initial positions.

This isn't a huge deal for, say, a grass tile which doesn't move, but
any moving tile might move around with the camera, and particularly,
the player at the moment *always* has the camera centered on them. If
the camera moved far enough that the player's original position was
no longer in the camera's bounds, the player would stop rendering!

This wasn't super obvious, because as long as the player was within the
bounds, the player would render in the *correct position*, because once
they are gotten from the quadtree, their own coordinates are used to place
them on the screen, which were correct. The fix here (which is kludgey and
needs to be made more smooth) is to directly inform the entity manager in
each object that moves' update methods that it needs to be plucked from
the quadtree (at the moment using the very slow 'get_by_id' quadtree call),
and then reinserted at their present coords.

This commit also includes a couple of small lint-fixes -- a semicolon which
should have been a comma, some comments moved around, some old console
logs being deleted, etc.