Add camera manager.
This adds a basic camera to the game that (for now) always centers
on the player, allowing movement in and display of maps that are
larger than the rendered screen.

The camera has a concept of a margin built-in so that the rendering
quadtree can extend a little bit past the screen, which allows objects
to continue drawing when only partially visible. That is a bit of a
hack at the moment, hardcoded to 100 pixels around the edges -- any
sufficiently large tile will still have some odd pop-in and a better
fix may need to be written someday.

The camera supports direct move and resize calls, but also has support
to be centered on a point, which is a pretty useful default behaviour.

It'd be nice to implement some clamping on the map bounaries, easing to
make it less jerky with player movement, or a collision-zone that causes
movement, but those may end up being left as the concerns of the users
of the camera, rather than the camera itself. A problem for another day.