Override player's layer to map's player_layer.
The player should be where the map tells 'em to be. This is *very*
kludgey at the moment but needs to be handled as I get into making
more complex laps with objects drawn both over, under, and at the same
height as the player.

The "right fix" for this is going to be to set the player's layer
during map load as part of the map load actions work I have planned.