Allow tiles to specify source data.
Tiles should not have to specify a source x/y/width/height, and the
resources have those defined as 0/their width/height if tiles do not
carry the information with them, but if a tile is only a piece of a
composite resource, rather than muddy up the resource map with copies
of the same resource that have different sources, this information can
be left up to the tile.

The code will now check for and prioritize the tile's information over
the resource information, and use whichever it ends up with.