Add some more complex test maps.
These are not totally ideal maps for this engine: the tiles are small and
square and repeat and overlap. Still though, they're a lot more of
an interesting thing to hop around in than the previous ones!

Exported from the 'editor' project I have hosted in my github at
https://github.com/wcarss/editor, which I've made a few small tweaks
to today in aid of creek-export support.