Add game and map init/deinit and update hooks.
A game's config doesn't have to define init or update hooks for
the game or any maps in the game, but doing so will cause them to
be called at game start, on map load, on map unload, and at each
frame.

In general, a game's updates should be cross-map global concerns,
while a map's updates should only be on that map.

Partly violating that suggestion, this commit also moves the
map-change and particle-spawn controls previously hard-coded into
the entity manager into the config's game-update definition to test
that the code continues to work. The map-change code should definitely
be part of the game-update definition, but as the particles are added
*to the map*, and get lost on each map change, that control should
be defined in the map's update.

I may need to make a distinction between "persistent objects" like the
player that need to be transferred between maps and others, like tiles
or these particles, that should be discarded. The concept of changing
the map you're interacting with so that it remains different when you
return to it is also one I should tackle -- I think though, you could
likely more or less handle that with the init/deinit code. Needs some
further looking into it.

This commit also has a few very small fixes, like removal of console.logs
or addition of init messages.