Reduce ResourceManager indentation and complexity.
This is a somewhat complex little piece of code and is the only really
asynchronous part of the program, piling up some promises and then group-
resolving them, attaching listeners to new image elements, etc -- it's
doing a lot.

It was structured in such a way that there were functions nested inside
of functions and those inner functions had for loops and ifs and, well,
it's a little simpler to read now, hopefully. :)