Add simple text-overlay engine.
This is useful for debugging and potentially in-game stuff -- text
can define its own update() function so that it is live, and it gets
the frame-delta and entity_manager every frame so it has access to
most things of interest in the game.

A good improvement would probably be to allow drawing a box under the
text if specified, because the contrast is typically pretty low.