Default to fullscreen and support resizing.
Some of the code here is a bit of a shit show. The CameraManager
asks the ContextManager literally every time it's moved for the width
and height of the current context. There are definitely better ways to
do that, but they can wait for a day when I'm feeling more patient.

The RenderManager also asks the context manager for the latest context
every frame -- again, this feels backwards. The context manager should
alert these two that the context has been updated and they should both
fetch it just once, and otherwise rely on their cached information. But
right now, they don't, and it's fine.

The resize() function in the context manager is uglyrifically playing
double duty as a handler to the window.onresize event AND being a means
for the context manager to manually specify a size for itself -- again,
something I might fix later.

There's a 4-pixel height-fudge-factor that I think is chrome-specific, and
if you do 3 pixels or less it'll cause a scrollbar which will then cause a
horizontal scrollbar because the appearance of the first scrollbar does
not trigger a window resize event. So the fudge-factor remains and you can
see a small line of white at the bottom of the screen.

Also, this doesn't use the real fullscreen API that it likely should use.
My vague plan is to provide some kind of hover-button or menu-button when
I've implemented the UI Manager which can cause switching into real
fullscreen or not.

But in general: the camera, context, and renderer now all size and draw
things properly across the whole screen, and the demo page is constructed
to do so by default. I'll have to find a way to show instructions later.