First pass at supporting user-defined buttons.
This introduces a UIManager and has several changes in the ControlManager,
for example introducing "down_at", "down", "up_at" fields not just for the
new button types but for mouse controls and for keys as well. It has
complicated checking a key's state (you now have to check the key AND that
the key is down -- which I really don't like), so I'll be experimenting
with ways to slim that down.

The UIManager is responsible for the real meat of how the buttons work,
which right now is as HTML divs with click/mouseenter/mouseout events
registered. Those events just send signals to the ControlManager, and
then any bit of code with a ControlManager reference can poll to check
the state of any button it's interested in.

The ControlManager just does the record-keeping of the up/down/at state,
and doesn't pay attention (much) to what sort of object it's keeping
track of these things for, or (most importantly) how those objects work.

And it works. The buttons are ugly and positioning/styling them is a huge
pain in the butt, but when they are clicked they do things. This commit
has a demo button 'Next Map' that is now the control for cycling maps.

I am going to have to experiment in the future with:
 - not using html buttons because they are annoying to style
 - using referential positioning schemes rather than absolute coords
   - this is for flexible screen-size support which is the very next
     thing being added; imagine defining a button on "the bottom of
     the screen" when the screen can be any height
 - slimming down the control interface while maintaining richness &
   consistency
 - configuration of buttons and controls

But for now, the basics are there. :)