Fix canvas "height-fudge" using display: block.
I've long wondered "what the deal is" with regard to 100% height
canvases having a scrollbar appear -- it turns out: canvas is an
inline element by default! Inline elements (if you're not aware)
have some complicated logic to them for determining their line-height,
and given their default text-alignment, line-height, and font-size,
typically extend a few pixels below the bottom of the content box --
this is to allow descenders (the bits of a font that drop below the
baseline of the font, like the tail of a y) to render inside of the
inline element.

So a 100% canvas, by default, is 100% _plus a few pixels_, to allow
for descenders to be rendered within it. But if I'm not putting any
text there... there's no need!

I fixed this by using display: block, which eliminates the complex
calculation of the inline box and the descender space, but there
are many possible solutions here -- setting display to absolute
(which takes the element out of the flow of the page), setting the
font-size and line-height of the canvas and possibly some of its
ancestors to 0 (which should make the inline-box calculation result
in 0px allocated for descenders), setting the vertical-align of text
to the top (which means descenders should no longer drop below the
bottom of the box), etc.

Neat.