Significantly improve bounded camera-centering.
The previous implementation made some effort to be clever by storing
the bounding rectangle for camera movement in the corner of the camera
and trying to assess the bounded movement conditions after transforming
the coordinates to that space -- it was not a helpful layer of
abstraction.

This commit puts a box centered inside the camera, half its width and
half its height, and works with the center-oriented coordinates that
are actually used by the engine to move the player around. It checks
if the player is outside of that bounding box in each dimension, and
if so, it calculates the resulting next-camera position in that
dimension.

A move is then issued to whatever coordinates have been calculated,
which, if the player was inside of the centered box, are the same as
the coordinates it had before the center call was issued.

It may not be the smoothest scrolling implementation -- I had some
thoughts on possibly causing being outside of the box to apply an
acceleration and to give the camera physics properties like velocity,
then to just update it each tick like the player. This could allow
for some interesting dynamic camera behaviour and I still might try
it, but the central bounding box is all I want to put time into for
now.

This commit also expands the field test map with dummy markers to
be meaningfully larger than my screen in all directions, for the
purposes of manually testing the bounding code. Going to have to
make some for-real larger test maps soon. :)