Re-add map-cycle key and move player start.
Hitting the 'Next Map' button has proven sufficiently onerous to re-add
the 'm' hotkey for cycling maps.

The player was starting in the very top-left corner and it was kind
of a crappy place to start -- they ought to have some sort of map-
specific starting position anyway, but for now they'll be out in the
world a little.