Improve ControlManager interface.
Checking that a Key is set and *also* that the Key is down is, well,
infuriating. It was terrible. But it enabled such a rich set of data!

... so, I'm keeping that interface along with its up/down_at stuff,
but it's being shuffled off into the background like it ought to be,
there only for anyone who might actually want it.

The New Front Doors to the ControlManager are the .keys, .buttons, and
.mouse methods, which will check both that the relevant id (if applicable)
exists and its down property is set. This makes all such checks easy on
the eyes, and doesn't interfere with the "rich" data in the background.