Add RequestManager for XMLHttpRequests.
This app is built more on polling than on promises, so the structure
of this manager is such that an id, if not given, will be assigned and
the request can be checked/live-updated later on, rather than such that
you would pass a button or something in a callback to the request's
success function. You make the request, it'l tell you an identifier, and
then you can ask later using that identifier if the request is done and
how it went.

This commit also contains some test-stuff for the RequestManager -- the
test.json file, the addition of a dummy display of loaded text from that
file in the game window, and a second request made right after to verify
that ids will be assigned reasonably.

The ids are time-based and (for now) pretty long and unreadable. I would
likely rather make a uniqueish memorable alphabetical hash for them, but
that's silly monkeying around rather than valuable work. They go to three
digits of the microsecond precision that performance.now() outputs, so
two executions side-by-side should have distinct ids.