Expose UI Manager through Entity Manager.
There are a lot of places this is sort of required, so I'm surprised
I hadn't done it yet. It might also be gud to add convenience methods
to the Entity Manager itself that add/remove UI directly and/or assume
sane defaults to ease UI management, but that can be thought out later.