Better default for map manager last-load time.
Setting it to "now" even though we didn't do a change means that if
we DO initiate a change to call a map's init() function right away
in the game start code, our legit call gets hammered. No bueno.

So setting it to 0 means the first call to a map change will go through.

:ok_hand: