Expose 'remove_entity' and add clear_all mock.
The clear_all mock may be filled out someday -- maps automatically
reset when a map_change is issued, to just what they had listed in
their layers in the config at game-boot. So entities are thrown away
at the point they are discovered to be stale and they are re-set-up.

But at some point I'm going to need to do persistence of entities
between maps that goes beyond 'write them to the layers in the config',
probably, so I'll also need a way to force them all to go away? Likely.

This also exposes the 'remove_entity' method, which I hadn't needed
until now I guess!