Add empty_dict checker to utilities.
Checking if a dictionary has no keys is a surprisingly tricky task.
{} isn't just falsey, for example, and {} has no length attribute.
The most common suggestion I saw was to check the Object.keys method's
length, which is awful! But it's what I've gotta do. Rather than drop
that little gem everywhere, it makes sense to hide behind a utility.