Totally revamp manager infrastructure.
These changes make it so that managers init each other as required,
eliminating the need for the complex dependency-tracking I'd been
doing as I added more or made modifications to their privileges. The
previous pattern had also required a _lot_ of argument passing and
modifications across multiple interfaces, which this eases pretty
substantially.

Also, most managers had previously been "gotten" via the EntityManager,
and that wasn't really its job -- also leading to the EntityManager
being injected into a lot of update() and init() functions throughout
the code; the pure Manager is now responsible for most of this.

This change broke a lot of things initially when I prototyped it in
the Tetris project, but things have stabilised there, and I think it's
more or less safe to backport into the main creek project now.