Strict-mode fixes.
I think these are (chronologically) an earlier set of changes than some
of the commits in the tree that have been placed prior to them, but most
of this work got done when I was hunting for collision-related bugs, so
this is a decent spot for it to land.

I should have been using strict mode all along for this project, and I
didn't actually know at the outset of the project that the iterator
in a for loop's implicit declaration is var-based and so leaks scope
all over the code. Using strict mode forces you to declare those, among
many other little things, and it really tightened the ship.