Move initialization of entitymanager's texts.
This feels like a hacky thing to do, and it is, but it's also the
correct thing to do. If the game's init method wanted to add texts,
it couldn't, because the game's init method is executed inside of the
entity manager's init method, so the texts variable wasn't initialized
to be added to, because the texts variable was initialized later in the
entity manager's init method than its game state variable.

It's not incredibly common to add text from the game's init method --
in this case it is very much debug text -- but that shouldn't be an
impossibility. Moving the text initialization up higher, as it has
no dependencies, just seems reasonable, but it does point to the
possibility of issues like this arising again in the future.