Use dynamic script loading to load maps.
The maps half been being loaded out of JSON files, which meant that
they couldn't host their own code, and I had to leave things in a
more or less half-done state with regard to the use of the map-files,
because I want the maps to host their own code, but I knew this would
be a fair bit of work to make possible.

On the plus side, I now have a script manager *and* a request manager,
and they both seem to work well.