Fix map inits: no more stale entities!
On the first map to load, if entities were dynamically added in the
map's init() method, they would be wiped out by the stale_entities()
system instantly. There were a lot of ways to fix this, but the most
clear-cut and fun was: wipe out the stale_entities system instantly.

The system was developed to limit bi-directional communication between
modules, which I still want to limit, but it was built on the fear
that such communication would be rampant and hard to think about,
which it is not (.. yet). Trying to "magically refresh entities" is
far more convoluted than a single call back to the entity manager
from within the map manager -- at least, it is now. At the time the
stale entities code was written, it would have been pretty bad, but
the manager() overhaul made limited access to a manager easy to get,
and here we are.

There's some improvements to is_loading and get_entities and other
stuff that needed to come along with this to prevent code from trying
to read too deeply out of an unloaded map.