Prevent double-initialization of maps.
Due to some honky code I ended up with change_maps(x) being called
twice in a project, and while that code got removed, it shouldn't
have been allowed to initialize the same whole map twice in a row.

It's possible there's a reason to do that, but if it is, I'll find a
way to explicitly allow it, rather than to allow it by default.