Implement some entity-lookup caching.
In the draw loop, we can have multiple people end up asking for the
currently-visible set of entities, and a tree-scan is a little bit
expensive, so we'll just hand you what we got last time if you ask
again too quickly.

This caching is a bit janky, and I might find a better way to achieve
similar results in the future. The number definitely needs tuning.

But for now, it speeds things up nice.