Only collide nearby objects.
This is ostensibly what a quadtree ought to be used for! Fewer things
to collide against things means lightning speeeeeed!

One slight break in this though is that I'm storing objects in the
tree by their x-y coords, not by their rect-positions.

So Let's say the hero is on top of (and trying to collide with) the
state of Texas. Well, the state of Texas' x-y coords may be nowhere
near the quadtree-leaf the hero's sitting in, nor any of its siblings'
siblings' siblings, and the hero (being a small-fry) would not get the
code to look around far enough to find them, despite the lone star
state's vast width and height encompassing all the hero can see.

In other words, big objects won't collide with small objects at the
edge of the big object. They'll collide at (roughly) the edge of the
small object.

I'm willing to wait until I have that problem to solve it. I could end
up rewriting the majority of the quadtree or discarding it by then
anyway, so investing in a solution to it today feels like a waste of
time.