Add index on entity-id to quadtree.
The quadtree_get_by_id method before now required a _full tree scan_,
which in a tree of O(1000) entities was a terrible idea to do inside
a draw-loop, let alone a tree of O(10k) or more entities.

What's worse, removal by id by used in move, so in many frames, we
could have dozens of calls to scan the whole tree for an entity
to remove -- it was awful.

Adding an index on id means keeping one global object of all entity
ids in the entire tree, and giving every subtree/leaf access to read
and write from it. When a node is inserted or removed, it is added to
or removed from the index as well in an O(1) operation, and this makes
get by id O(1), even better than get by coord at O(logN).

Removal has a little hitch that makes it not quite so easy: we have to
find the list of entities belonging to the specific leaf node that a
given entity is stored in in order to splice the entity out of that
list. Rather than require an O(N) tree scan, or get-by-id and do an
O(logN) remove-by-coords w/ id constraint, we keep a bookkeeping
pointer to eachh entity's local list as a property on each entity,
so that with a given entity, you can remove it from the tree in O(M),
where M is number of entities in a specific leaf of the tree (in
practice a very small number).

This solution (a new key on the objects being stored) carries the risk
of having the key be in-use by app code before or after the quadtree
sets it. Because of that, I would rather store another piece of
metadata internal to the tree, but in practice (as I'm the only user
here), it's not that big a deal. At some point I will fix it.