Reduce entity move/remove conflicts.
By letting the quadtree code handle movement within the tree, we can
peer into details like whether the entity was removed from the index
(indicating we want it to go away) or was left (indicating this is
just a movement from one place in the tree to another), and can do
some bookkeeping on our own actions as well.

This solves a bug where an entity would undergo movement at the same
time it would be removed, resulting in a race that could re-insert a
removed entity, and a host of downstream issues from there. Looking
into the index before re-inserting in a move and checking some lock-
like bookkeeping values for movement and removal should make these
operations safer.

It's no guarantee (as the operation is totally not atomic!) but it
would be a lot harder now for an object to get into a funky state.