Reverse effect of 'Only collide nearby objects'.
I hit a case where this was a problem much faster than I thought I
would: updating the creek engine in tetris instantly broke, because
I use large rectangular bound objects to collide against the smaller
square pieces that make up a tetris shape to determine left, right,
and bottom stopping conditions in that game, and the local entities
optimization broke those collisions.

This wasn't a total dealbreaker in jabiru, just a modest gain, so
we can hopefully leave this as is until I find time to improve the
quadtree (or find a suitable replacement) to deal with that type of
case.