Use local canvases as resource sources.
Drawing tiles out of big images and scaling them to a particular size
on every frame seems to be asking a lot of the computer. If we could
just have the images be on their own and stored at a size we plan to
make direct use of, there would be no extra work for the computer to
do each frame. And we can! This code does.

If a tile has a scale factor that makes it the same size as its stored
resource and the resource has source x/y coordinates of 0,0, creek will
now use the simplest drawImage function that applies no transformations,
and that (in my unscientific observations of repeated reloads and
watching the dev tools FPS metre / analysing drawImage times in perf
data captures) performs much better than its more argument-heavy
sibling.

Generally, tiles aren't changing sizes and resources aren't changing
their sources too often, and the extra checks per frame don't seem to
eat any meaningful time.