Add a version file.
The version we're going to pretend we're starting at here is 0.5.0,
and the pattern of version changes will be roughly semver: in x.y.z
versions z will increment for typos and minor patches and bugfixes
that do not break compatibility, y will increment for new additive
features or otherwise larger changes that retain compatibility, and
x will increment for any feature that breaks compatibility.

This is necessary mostly to start keeping track of what "version" of
Creek is in different projects, rather than just copying it back and
forth at my convenience and losing track.