Add Controller Manager.
This is not for physical joypads (at least, at present) -- it is for
virtual controllers displayed on-screen to support better mobile play
experiences. Touch controls are too clunky to achieve many different
actions intuitively and reliably with gestures alone, so a classic
button-controller scheme is my best idea for a fallback.

I intend to allow the user to define controllers for their own use,
and to put their specifications in a different file, etc, but for now
this is a working portrait and landscape NES-compatible controller
that creates buttons via the UI manager.

One major change that remains is to offset+limit the canvas+camera size
to respect the controller, and keep the rest of the game on-screen.

The landscape-controller code is currently enabled-by-default in the
sample project, while surrounding work is in-progress.