Resize+move canvas+camera for Controllers.
The controllers take up a lot of screen real estate -- rather than
have the game render behind them, we can adjust things.

UI elements that are drawn on top of the game should be offset by the
size of the controller above/to the left of them. I'm not handling the
case of buttons on the right being pushed rightward yet, but will have
to soon.

The canvas + camera itself now shrinks and is pushed as necessary to
get out of the controllers' way, and this is handled correctly across
changes of controller type and removal/re-addition of controller UI,
so that a user changing from portrait to landscape and back on a phone
will not end up with a broken experience / can have the control scheme
they prefer.

The sample project now starts without the virtual controller active,
and includes a 'Controller' button that activates the portrait-mode
controller, then cycles to the landscape, and then deactivates the
virtual controller. The START button on the controller also cycles
to the next controller, and the buttons/d-pad are hooked up to various
sample functions.