Clear screen on map-change & texts on removal.
If you don't want to clear the whole screen every frame, but you
draw every frame, and you don't draw anything overtop of your text,
you end up repeatedly overdrawing your text.

If you remove your text, you no longer have anything to draw, so you
don't clear the text! So this adds a manual clear step at the point
of removal for texts. It unfortunately clears the *whole* screen,
which is expensive, and looks bad. For that reason and the overdrawing
mentioned in the first paragraph, I'll have to change the text render
model in the future.

For now though, manually clearing the screen on removal is fine, and
manually clearing the screen at map-change is just good hygiene.